+++ resources/views/charts/flux/line.blade.php
@@ -1,8 +1,0 @@
-@php $series = $data['series'] ?? []; @endphp
-<flux:chart class="aspect-[3/1]">
-    <flux:chart.svg>
-        @foreach($series as $s)
-            <flux:chart.line :field="@json($s['name'] ?? 'value')" />
-        @endforeach
-    </flux:chart.svg>
-</flux:chart>